+++ app/Http/Controllers/PostController.php
@@ -8,6 +8,7 @@ use App\Models\Post;
 use App\Http\Resources\PostCollection;
 use App\Http\Resources\PostResource;
 use App\Models\Topic;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class PostController extends Controller
@@ -31,7 +32,14 @@ class PostController extends Controller
             ->setMessage("Post retrieved successfully");
     }
 
-    public function store(Request $request) {
+    public function store(Request $request, int $id) {
+        $topic = Topic::find($id);
+
+        if (!$topic) throw new NotFoundException("Not found", json_encode(["not_found"=> "The topic you're trying to access does not exist"]));
+
+        //  Check user permissions to create post
+        $this->authorize("create", [Post::class, $topic]);
+
 
         $validator = Validator::make($request->all(), [
             "post_title"=> "required|string",
@@ -57,8 +65,8 @@ class PostController extends Controller
             "post_content"=> $request->get("post_content"),
             "post_coordinates"=> serialize(array("lat"=> $request->get("post_coordinates_lat"), "long"=> $request->get("post_coordinates_long"))),
             "post_type"=> "post",
-            "post_author"=> $request->get("post_author"),
-            "post_parent"=> $request->get("post_parent")
+            "post_author"=> Auth::id(),
+            "post_parent"=> $topic->id
         );
 
         $post = Post::create($data);
@@ -70,7 +78,10 @@ class PostController extends Controller
             ->setMessage("Post created successfully");
     }
 
-    public function update(Request $request, $id) {
+    public function update(Request $request, int $id) {
+        $old_post = Post::findOrFail($id);
+
+        $this->authorize("update", [$old_post]);
 
         $validator = Validator::make($request->all(), [
             "post_title"=> "string|nullable",
@@ -126,10 +137,9 @@ class PostController extends Controller
 
         $post = Post::find($id);
 
-        if (!$post) return BaseResource::error()
-            ->setCode(404)
-            ->setMessage("Post not found")
-            ->setErrors(json_encode(["not_found"=> "The post you're trying to delete does not exist"]));
+        if (!$post) throw new NotFoundException("Not Found", json_encode(["not_found"=> "The post you're trying to delete does not exist"]));
+
+        $this->authorize("delete", [$post]);
 
         $post->delete();
